+++ app/controllers/document_exports_controller_base.rb
@@ -88,6 +88,7 @@ class DocumentExportsControllerBase < ApplicationController
       'HudApr::DocumentExports::HudAprExport',
       'HudApr::DocumentExports::HudCaperExport',
       'HudApr::DocumentExports::HudCeAprExport',
+      'HudApr::DocumentExports::HudDqExport',
       'HudPathReport::DocumentExports::HudPathReportExport',
       'HudSpmReport::DocumentExports::HudSpmReportExport',
       'HudDataQualityReport::DocumentExports::HudDataQualityReportExport',
