+++ drivers/hud_apr/app/controllers/hud_apr/dq/questions_controller.rb
@@ -1,0 +1,18 @@
+###
+# Copyright 2016 - 2023 Green River Data Analysis, LLC
+#
+# License detail: https://github.com/greenriver/hmis-warehouse/blob/production/LICENSE.md
+###
+
+module HudApr::Dq
+  class QuestionsController < HudApr::QuestionsController
+    include DqConcern
+    before_action :set_report, only: [:show, :destroy, :running, :result]
+    before_action :set_question
+    before_action :set_reports
+
+    private def report_param_name
+      :dq_id
+    end
+  end
+end
