+++ drivers/hud_apr/app/controllers/hud_apr/dqs_controller.rb
@@ -1,0 +1,15 @@
+###
+# Copyright 2016 - 2023 Green River Data Analysis, LLC
+#
+# License detail: https://github.com/greenriver/hmis-warehouse/blob/production/LICENSE.md
+###
+
+module HudApr
+  class DqsController < BaseController
+    include Dq::DqConcern
+    before_action :generator, only: [:download]
+    before_action :set_report, only: [:show, :destroy, :running, :download]
+    before_action :set_reports, except: [:index, :running_all_questions]
+    before_action :set_pdf_export, only: [:show, :download]
+  end
+end
