+++ drivers/hud_apr/app/models/hud_apr/document_exports/hud_dq_export.rb
@@ -4,23 +4,22 @@
 # License detail: https://github.com/greenriver/hmis-warehouse/blob/production/LICENSE.md
 ###
 
-module HudDataQualityReport::DocumentExports
-  class HudDataQualityReportExport < ::GrdaWarehouse::DocumentExport
+module HudApr::DocumentExports
+  class HudDqExport < ::GrdaWarehouse::DocumentExport
     include ApplicationHelper
     include HudReports::HudPdfExportConcern
 
     def generator_url
-      hud_reports_past_dq_path(report)
+      hud_reports_dq_path(report)
     end
 
     private def controller_class
-      HudDataQualityReport::DqsController
+      HudApr::DqsController
     end
 
     private def possible_generator_classes
       [
-        HudDataQualityReport::Generators::Fy2020::Generator,
-        HudDataQualityReport::Generators::Fy2022::Generator,
+        HudApr::Generators::Dq::Fy2024::Generator,
       ]
     end
   end
