+++ drivers/hud_apr/app/models/hud_apr/generators/dq/fy2024/base.rb
@@ -1,0 +1,10 @@
+###
+# Copyright 2016 - 2023 Green River Data Analysis, LLC
+#
+# License detail: https://github.com/greenriver/hmis-warehouse/blob/production/LICENSE.md
+###
+
+module HudApr::Generators::Dq::Fy2024
+  class Base < ::HudApr::Generators::Shared::Fy2024::Base
+  end
+end
