+++ drivers/hud_apr/app/models/hud_apr/generators/shared/fy2024/question_four.rb
@@ -8,25 +8,27 @@ module HudApr::Generators::Shared::Fy2024
   class QuestionFour < Base
     QUESTION_NUMBER = 'Question 4'.freeze
 
-    TABLE_HEADER = [
-      'Organization Name',
-      'Organization ID',
-      'Project Name',
-      'Project ID',
-      'HMIS Project Type',
-      'RRH Subtype',
-      'Coordinated Entry Access Point',
-      'Affiliated with a residential project',
-      'Project IDs of affiliations',
-      'CoC Number',
-      'Geocode',
-      'Victim Service Provider',
-      'HMIS Software Name',
-      'Report Start Date',
-      'Report End Date',
-      'Total Active Clients',
-      'Total Active Households',
-    ].freeze
+    def table_header
+      [
+        'Organization Name',
+        'Organization ID',
+        'Project Name',
+        'Project ID',
+        'HMIS Project Type',
+        'RRH Subtype',
+        'Coordinated Entry Access Point',
+        'Affiliated with a residential project',
+        'Project IDs of affiliations',
+        'CoC Number',
+        'Geocode',
+        'Victim Service Provider',
+        'HMIS Software Name',
+        'Report Start Date',
+        'Report End Date',
+        'Total Active Clients',
+        'Total Active Households',
+      ].freeze
+    end
 
     HMIS_SOFTWARE_NAME = 'OpenPath HMIS Warehouse'.freeze
 
@@ -52,7 +54,7 @@ module HudApr::Generators::Shared::Fy2024
 
     private def q4_project_identifiers
       @report.universe(QUESTION_NUMBER)
-      headers = TABLE_HEADER.zip(('A'..'Q').to_a)
+      headers = table_header.zip(('A'..'Q').to_a)
 
       question_sheet(question: 'Q4a') do |sheet|
         headers.each { |label, col| sheet.add_header(col: col, label: label) }
