+++ drivers/hud_apr/app/views/hud_apr/dq/cells/show.haml
@@ -1,0 +1,1 @@
+= render 'hud_apr/shared/cell_details'
