+++ drivers/hud_apr/app/views/hud_apr/dq/cells/show.xlsx.axlsx
@@ -1,0 +1,12 @@
+wb = xlsx_package.workbook
+wb.add_worksheet(name: @name.slice(0,30).gsub(':', ' ')) do |sheet|
+  title = sheet.styles.add_style(sz: 12, b: true, alignment: {horizontal: :center})
+  sheet.add_row(HudApr::Fy2020::AprClient.detail_headers_for_export.values, style: title)
+  @clients.each do |client|
+    row = []
+    HudApr::Fy2020::AprClient.detail_headers_for_export.each_key do |k|
+      row << client[k]
+    end
+    sheet.add_row(row)
+  end
+end
