+++ drivers/hud_apr/app/views/hud_apr/dqs/_breadcrumbs.haml
@@ -1,3 +1,3 @@
 = content_for :crumbs do
-  = link_to history_hud_reports_dqs_path do
+  = link_to hud_reports_dqs_path do
     = "« HMIS Data Quality Reports"
