+++ drivers/hud_apr/app/views/hud_apr/dqs/history.haml
@@ -1,0 +1,1 @@
+= render 'hud_reports/history'
