+++ drivers/hud_apr/app/views/hud_apr/dqs/index.haml
@@ -1,0 +1,1 @@
+= render 'hud_reports/index'
