+++ drivers/hud_apr/app/views/hud_apr/dqs/running.js.coffee
@@ -1,0 +1,1 @@
+$('.jReportListing').html "<%=j render 'hud_reports/reports' %>"
