+++ drivers/hud_apr/app/views/hud_apr/dqs/running_all_questions.js.coffee
@@ -1,0 +1,1 @@
+$('.jReportListing').html "<%=j render 'all_questions' %>"
