+++ drivers/hud_apr/app/views/hud_apr/dqs/show.haml
@@ -1,0 +1,1 @@
+= render 'hud_reports/show'
