+++ drivers/hud_apr/app/views/hud_apr/questions/show.haml
@@ -8,6 +8,4 @@
 
 = render 'hud_apr/shared/report_form', url: path_for_questions(@question)
 
-- if @reports.present?
-  = render 'hud_reports/year_tabs'
 = render 'hud_reports/reports_wrapper'
