+++ drivers/hud_data_quality_report/app/controllers/hud_data_quality_report/questions_controller.rb
@@ -33,7 +33,7 @@ module HudDataQualityReport
     end
 
     private def report_param_name
-      :dq_id
+      :past_dq_id
     end
   end
 end
