+++ drivers/hud_data_quality_report/app/models/hud_data_quality_report/generators/fy2020/generator.rb
@@ -19,7 +19,7 @@ module HudDataQualityReport::Generators::Fy2020
     end
 
     def url
-      hud_reports_dq_url(report, { host: ENV['FQDN'], protocol: 'https' })
+      hud_reports_past_dq_url(report, { host: ENV['FQDN'], protocol: 'https' })
     end
 
     def self.questions
