+++ drivers/hud_data_quality_report/app/views/hud_data_quality_report/dqs/_all_questions.haml
@@ -1,20 +1,16 @@
 .well
   %p
     The Data Quality Report contains 7 questions, labeled 1 through 7.  Each question contains a single table.
-    Questions can be run individually below, or can be run all together. Each report run can be downloaded as a
-    zip file suitable for upload to HUD by following the link in the
+    Each prior report run can be downloaded as a zip file suitable for upload to HUD by following the link in the
     %strong Status
     column on subsequent pages.
   %p
-    = link_to history_hud_reports_dqs_path do
+    = link_to history_hud_reports_past_dqs_path do
       See the history of all report runs and download zip files
-  %p
-    = link_to new_hud_reports_dq_path do
-      Generate New Data Quality Report
 
 = render 'hud_reports/all_questions_table'
 
 .well
   %p
     Previously run FY2017 Data Quality Reports can be accessed from
-    = link_to 'here.',  hud_reports_legacy_dqs_path
+    = link_to 'here.',  hud_reports_legacy_past_dqs_path
