+++ drivers/hud_data_quality_report/app/views/hud_data_quality_report/legacy_dqs/index.haml
@@ -1,5 +1,5 @@
 = content_for :crumbs do
-  = link_to hud_reports_dqs_path do
+  = link_to hud_reports_past_dqs_path do
     = "« Data Quality Reports"
 
 %h1 Legacy Data Quality Reports
@@ -18,9 +18,9 @@
   %tbody
     - @reports.each do |report|
       %tr
-        %td= link_to report.name, hud_reports_legacy_dq_path(report)
+        %td= link_to report.name, hud_reports_legacy_past_dq_path(report)
         %td= report.last_result(current_user).try(:last_run)
         %td
           - if status = report.last_result(current_user).try(:last_run_status)
-            = link_to status, hud_reports_legacy_dq_legacy_result_path(report, report.last_result(current_user))
+            = link_to status, hud_reports_legacy_past_dq_legacy_result_path(report, report.last_result(current_user))
         %td= report.last_result(current_user).try(:completed_in)
