+++ drivers/hud_data_quality_report/config/initializers/hud_data_quality_report_feature.rb
@@ -11,13 +11,3 @@
 #
 # use with caution!
 RailsDrivers.loaded << :hud_data_quality_report
-
-Rails.application.config.hud_reports['HudDataQualityReport::Generators::Fy2020::Generator'] = {
-  title: 'Data Quality Report',
-  helper: 'hud_reports_dqs_path',
-}
-
-Rails.application.config.hud_reports['HudDataQualityReport::Generators::Fy2022::Generator'] = {
-  title: 'Data Quality Report',
-  helper: 'hud_reports_dqs_path',
-}
