+++ drivers/hud_data_quality_report/config/routes.rb
@@ -6,7 +6,7 @@
 
 BostonHmis::Application.routes.draw do
   scope module: :hud_data_quality_report, path: :hud_reports, as: :hud_reports do
-    resources :dqs do
+    resources :past_dqs, controller: 'dqs' do
       get :running, on: :collection
       get :running_all_questions, on: :collection
       get :history, on: :collection
